+++ client/incoming-redirect/paths.js
@@ -1,0 +1,5 @@
+export const incomingRedirectRoot = '/incoming-redirect';
+
+export function emailManagementTitanExternal( mode ) {
+	return incomingRedirectRoot + '/titan/' + mode;
+}
