+++ client/my-sites/email/paths.js
@@ -100,7 +100,3 @@ export function emailManagementEdit( siteName, domainName, slug, relativeTo = nu
 export function isUnderEmailManagementAll( path ) {
 	return path?.startsWith( emailManagementAllSitesPrefix + '/' );
 }
-
-export function emailManagementTitanExternal( mode ) {
-	return '/titan/' + mode;
-}
